+++ src/Schema/Directives/RelationDirective.php
@@ -9,6 +9,7 @@ use Illuminate\Contracts\Config\Repository as ConfigRepository;
 use Illuminate\Database\ConnectionResolverInterface;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
+use Illuminate\Database\Eloquent\Relations\MorphTo;
 use Illuminate\Database\Eloquent\Relations\Relation;
 use Nuwave\Lighthouse\Exceptions\DefinitionException;
 use Nuwave\Lighthouse\Execution\BatchLoader\BatchLoaderRegistry;
@@ -54,15 +55,23 @@ abstract class RelationDirective extends BaseDirective implements FieldResolver
             // that we know to be present on the parent model.
             if (
                 $this->lighthouseConfig['shortcut_foreign_key_selection']
-                && (array_diff_assoc($resolveInfo->getFieldSelection(), ['id' => true, '__typename' => true]) === [])
+                && (array_diff(array_keys($resolveInfo->getFieldSelection()), ['id', '__typename']) === [])
                 && $relation instanceof BelongsTo
                 && $args === []
             ) {
                 $id = $parent->getAttribute($relation->getForeignKeyName());
 
-                return $id === null
-                    ? null
-                    : ['id' => $id];
+                if ($id === null) {
+                    return null;
+                }
+
+                // If the relation is polymorphic return the model instance so that TypeRegistry::typeResolverFallback can resolve the correct type
+                if ($relation instanceof MorphTo) {
+                    $model = $relation->getModel();
+                    $model->id = $id;
+                    return $model;
+                }
+                return ['id' => $id];
             }
 
             if (