+++ CHANGELOG.md
@@ -9,6 +9,10 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ## Unreleased
 
+### Fixed
+
+- Fix shortcut foreign key selection for morphTo relationships that use inline fragments in field selection https://github.com/nuwave/lighthouse/pull/2670
+
 ## v6.51.1
 
 ### Fixed